[FIX] planning: fix planning calendar test
Before this commit, a planning calendar test was wrong. In this test,
we resized two events to make them end at 14:30 but then check that they
end at 15:30 which is inconsistent.
This commit adds a mockDate with a defined timezone which will prevent
the test to depend on the machine local timezone.

closes odoo/enterprise#65599

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>