[IMP] *spreadsheet*: change layout of spreadsheet actions
With this commit, the layout of the spreadsheet actions is reworked to
have more space to display the spreadsheet.

The navbar and the control panel are now merged into a single line.

Technical details:
- The control panel is completely removed
- All spreadsheet actions are now in "fullscreen" mode
- All spreadsheet actions now have a custom navbar, extended from the
  web_enterprise navbar, in order to display:
    1. The spreadsheet icon
    2. The "app" name "Spreadsheet"
    3. A simplified breadcrumb => only the last part of the path
    4. The title of the spreadsheet
    5. (only for documents) The favorite toggler
- The actions that were in the control panel ("Share", "Filters", the
  save indicator, ...) are now moved inside the spreadsheet topbar. A
  component is created for each action in order to achieve it. The main
  technical impact of moving these actions is that they now should works
  with the `env` instead of `props`.

One known issue:
- Refreshing the browser while in a spreadsheet action will briefly
display the web navbar before the spreadsheet navbar is displayed. This
is due to the fact that the web navbar is displayed before the action is
loaded and therefore before the detection of the action mode `fullscreen`.

Task: 4000864
Part-of: odoo/enterprise#66209
Related: odoo/odoo#172202
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>