[FIX] mrp_workorder: add pager to shop floor
In [1], a limit was introduced to the amount of records that can be
displayed at the same time in the shop floor app. However, it was not
always very clear to the user that they may only be viewing a subset of
the available records. It was also impossible to show any records beyond
the first 40 that conform to the filter criteria.

In this commit, we introduce a pager to resolve these issues. The pager
will only be displayed when relevant (ie when not all available records
are displayed). Using this new pager it is also possible for the user to
dynamically increase or decrease the default limit of 40, which was not
possible before.

[1] https://github.com/odoo/enterprise/pull/64551 (and forward ports)

closes odoo/enterprise#65995

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>