[FIX] account_reports: print the correct ids in coverage reports
The coverage reports do not print the right account ids when searching for the ones that are not used.

Steps to reproduce:
- Install the 'l10n_de' module and switch to a German company
- Go in Accounting > Configuration > Management > Accounting Reports
- Go in the German Profit and Loss report
- Click on the button 'Accounts Coverage Report'
- In the downloaded report the yellow lines should not appear as they are linked

Cause:
When searching for the ids that are not linked, the code actually only look up the ones that are linked to an acocunt.

Solution:
Fix the condition on the ids.

opw-3993835

closes odoo/enterprise#65810

X-original-commit: 5a07c6a979c96a8b6890b40bfb4e4375c9e53fdd
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>