[IMP] documents: enable traceback sharing
This commit adds to the error_dialog the possibility to share the error
by creating a new text document on the fly in the automatically generated
support folder and then generating a share link for this document. A
new share button is therefore added to the error_dialog's footer as soon
as documents is installed and the generated share link is directly copied
upon pressing the button as well as displayed next to the button afterwards.

task-3883329

closes odoo/enterprise#65954

Related: odoo/odoo#170923
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>