[IMP] account_reports: use cog for buttons dropdown
Current behavior:
Currently, there is a dropdown button next to the first button visible in
account reports, where buttons like 'XLSX' and 'Save' are displayed. Buttons
with always_show are displayed next to the first button. Also, 'Save' is
displayed even when Documents is not installed.

Desired behavior:
- The buttons displayed inside the button dropdown should be moved to a cog
dropdown to the right of the title instead.
- 'XLSX' should be renamed to 'Download Excel' and 'Save' should be renamed to
'Copy to Documents'.
- 'Copy to Documents' should only be displayed if the module Documents is
installed.

Approach:
- A new component AccountReportCogMenu was created. The buttons bar only displays
the first button and always_show buttons (singleButtons) while the cog menu
displays the groupedButtons.
- 'Copy to Documents' is only appended to options['buttons'] in documents_account.
- The function open_report_export_wizard is moved to documents_account, since it
is not used elsewhere.
- Tests were also updated to reflect the change.

task-3986441

closes odoo/enterprise#65733

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>