[IMP] test_l10n_be_hr_payroll_account : allow user's to quick login
When several users share the same device, the login/logout process can
be boring, so we added a quick login feature.

For a security propose it doesn't store the session or the password of
the user. No data is sent to the server, everything is stored inside
the browser localStorage[1]. It only stores the login of the user and
the avatar of the user for the last 5 users connected to the device.

We also lightly redesign the login screen

task-3953707

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage

closes odoo/enterprise#66553

Related: odoo/odoo#172889
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>