[FIX] helpdesk_fsm: compute to correct default fsm project
Problem: Upon installing helpdesk_fsm in a multi-company setup,
the default fsm project when enabling "use_fsm" on helpdesk teams
is incorrectly retrieving the fsm project from the wrong company.
This leads to a multi-company access error when trying to navigate
to the fsm project linked to the helpdesk teams.

Purpose: Instead of incorrectly defaulting the fsm project from
the wrong company upon installation using a default method,
compute the default of fsm project based on the team's company
when use_fsm is enabled on the team.

Steps to Reproduce on Runbot:
1. Install Project, Field Service, and any l10n module
2. Switch to the localized company
3. Create a helpdesk team
4. Install the helpdesk_fsm module
5. Navigate to the helpdesk team settings
6. Enable "use_fsm" on the team
7. Try to access the default fsm project
8. multi-company AccessError is thrown

opw-3989799

closes odoo/enterprise#69577

X-original-commit: dc4eae41c69624a3006ffca8370f7cae3f16bb34
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>