[FIX] l10n_mx_edi: warn user when signing move for partner w/o country
When a contact is created with an RFC (VAT number) but no country
configured, signing an invoice for that contact results in the invoice
being signed correctly. However, the PDF shows the partner's RFC, while
the XML file shows the generic RFC (XAXX01010101). This discrepancy
misleads users into believing their setup is correct, as no error or
alert is provided.

This commit introduces a non-blocking warning message in the 'Send and
Print' wizard for cases where the contact's country is missing.

opw-3891896

closes odoo/enterprise#63255

Signed-off-by: Laurent Smet (las) <las@odoo.com>