[ADD] spreadsheet_sale_management: add quote calculator
Purpose
-------

In several businesses, it might be cumbersome to properly calculate a
quote (define an appropriate quantity, price, etc). The goal of this task is
to allow user to easily access a spreadsheet (blank or preconfigured) to
do those calculations.

Specification
-------------

From a quotation template, you can create a spreadsheet calculator template.
At creation, a list of sale order lines is inserted with a global
filter allowing to filter the lines according to their quotation.

You can add "Field sync" rules on any cell (right click on a cell) to update
one field of one order line in the list (readonly fields are excluded).

Then, when you create a quotation from the quotation template, you can open
the spreadsheet calculator. It creates a copy of the spreadsheet (allowing
you to play around values etc.) and adds a global filter to filter the
sale order lines from the quotation.

From this spreadsheet, by clicking on "Save in SO00XX", it reads the values
in cells with a "Field sync" rules and updates the quotation lines with
those values.

This allows to create very specific flows with any arbitraty computations
to compute any values in sale order lines, based on any data, from the
spreadsheet itself or from the database by inserting other data sources
in the spreadsheet template, for example, employee planning, products,
stock, etc.

This great power and flexibility comes with limitation about how generic
it can be. It's expected to hardcode product names or ids in formulas in
order to match order lines. Hardcoded names are of course not robust to
different languages, ids are robust to that but are less friendly and are
not cross-db compatible. Each spreadsheet calculator template must be
built specifically for every use case and every database.

Task: 3761590
Part-of: odoo/enterprise#65213
Related: odoo/odoo#170864
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>