[FIX] account_avatax: priority on address instead of latitude and longitude
Before this commit, avatax used the longitude and latitude in priority even if
the partner has an address.
The problem with this is that when you print the AvaTax report, the fields
DestStreet, DestCity, and DestPostalCode are not populated because AvaTax was
not sent any of that information. The customer needs those fields to be
populated or else they may be charged more money.

So now it will be the opposite, we will use the address in priority but if the
customer has a delivery address with no address information we will use the
latitude and longitude of the delivery address partner.

closes odoo/enterprise#66106

Task: 4011481
X-original-commit: a8b01faecf3b0c819397041ad8d4ac0ed9f10000
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>