[IMP] account_reports: create composite report from list view
The current flow to create a composite report is to create a new
report, toggle the "Composite Report" boolean field, and then set
the different reports in the list below. This task aims at slightly
improve the flow by also enabling the user to directly create composite
reports from the list view of reports. Users can select desired reports
to include in the composite group and initiate an action that generates
a new composite report. By default, the newly generated report will
include the selected reports as part of the composite group.

task-4019381

closes odoo/enterprise#65740

Signed-off-by: John Laterre (jol) <jol@odoo.com>