[FIX] sale_subscription: MRR currency rate
Steps to reproduce:
Create a Subscription order in a different currency (THS) than the main currency (USD).

Bug:
We consider that the MRR is a "living amount". The customer will continue to pay the amount is TSh.
This amount is converted into the currency of the company at the order date rate (from TSH to USD) which means,
the TSh value is fixed at order date which is not correct as customer will continue to pay in TSh and the rate evolve.

closes odoo/enterprise#61872

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>