[FIX] account_reports: previous in options
In this commit: https://github.com/odoo/enterprise/commit/ae9a269f640b848cffa79365a3091e6423f8e0dd
We removed a line setting the previous period in the options. But this line was
very useful for the export and tax closing. Since we don't go through the JS for
those two flows, the period was not set in the options and so the new_period is
by default -1.

Also correcting a type while renaming the field.

closes odoo/enterprise#65754

Task: 4019504
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>