[REV] web_enterprise: revert box-shadow var to 0
This reverts commit
odoo/enterprise@34a846cae4382b9507407e9f4abd509a02ce4331.

Setting the box-shadow to none will trigger a WARNING in the
compiler due to how bootstrap handle the box-shadow mixin[1] when
receiving two arguments with one containing a 'none'

We are setting the value back to 0 which will provide an invalid
property too when the mixin is called with a single argument but won't
warn in the console.

[1]: https://github.com/twbs/bootstrap/pull/30394

follow up of task-3755095

closes odoo/enterprise#65757

Signed-off-by: Pierre Paridans (app) <app@odoo.com>