[FIX] sale_{stock,mrp}_renting: return kit with subkit
When renting with deliveries, if it's about a kit with subkit and
shared component, the return will not be correct

To reproduce the issue:
1. In Settings, enable "Rental Transfers"
2. Create a kit K as follow:
   - K
     - C1
       - C2
       - C3
     - C2
3. Confirm a rental with 1 x K

Error: in the return, there are two lines for each C2, which makes
sense since each line is linked to a different kit. However, the
quantity on each line is incorrect: 2 instead of 1

When generating the SM of the return, we first generate the values
https://github.com/odoo/odoo/blob/8db471e74238481a5429f0fcf8fbce1c0960ff27/addons/stock/models/stock_rule.py#L205-L206
In `_push_prepare_move_copy_values`, at some point, we adapt the
quantity of the returned SM based on the other SM of the SO. To do
so, we only consider the SM with the same product (among other
conditions). This criterion is not strict enough: for instance, when
looking at C2 of K, we take its quantity *and* we also take C2 of C1,
which is incorrect. Of course, we do the same when looking at C2 of
C1, hence the incorrect quantities.

OPW-3996074

closes odoo/enterprise#65758

Signed-off-by: William Henrotin (whe) <whe@odoo.com>