[FIX] pos_settle_due: show settle due accounts button
The settle due accounts was not always shown on the partner screen.
Now it's always shown, and when the user doesn't have any amount due the
text changes to "Deposit Money" instead of "Settle Due Accounts"

Steps to reproduce:
-------------------
* Install pos_settle_due module
* Open PoS session
* Open partner list
* Click on details for any partner with no due
> Observation: The "Settle Due Accounts" button is missing

Why the fix:
------------
We always show the button, but change the text to better match the
behavior. If there is no amount due the text says "Deposit Money" and
when the user has some amount due it says "Settle Due Accounts"

opw-3925397

closes odoo/enterprise#66926

X-original-commit: 1582872ea6fa6fe7d995cb93affff650b6a9cff3
Related: odoo/odoo#173673
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>