[FIX] *: HOOT - deepCopy steps & matcher values
This commit adapts the enterprise code affected by a recent change:
`expect.step()` calls will deep copy their values, allowing the ensuing
`expect.verifySteps()` to check the value *as it was* when stepping it.
Some assertions were incorrect as they were checking the state of these
steps at the moment of the check, which is wrong.

X-original-commit: 0dad341508041b345aba4b462dcf11f36385d613
Part-of: odoo/enterprise#65963
Related: odoo/odoo#171752
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>