[IMP] *_barcodelookup: restrict barcode lookup call to EAN or UPC format
*: product, stock_barcode

Previously, the barcode lookup was triggered for barcodes of any length,
and the product creation form accepted barcodes of any length.

With this commit
=================
The barcode lookup will now be restricted to only those barcodes encoded
in EAN or UPC formats. Additionally, the product creation wizard will only
accept barcodes in the specified formats.

closes odoo/enterprise#66465

Task: 3974067
Signed-off-by: William Henrotin (whe) <whe@odoo.com>