[REF] web_studio: add support for widget options/attributes in Sidebar
This commit is part of the improvements allowing to edit some widgets directly in the
Sidebar, as well as fields were already available. Now, following commit (1), it is
possible to add documentation to supportedOptions and attributes on widget components.

This is directly used to display the options accordingly in the Sidebar.

(1): https://github.com/odoo/odoo/commit/097f6c8709e4cbecf8d08e3683225b76fde97f8f

closes odoo/enterprise#65792

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>