[FIX] l10n_ch_hr_payroll_elm: Fix Dependent children retrieval and payslip UX
When computing the number of children of an employee, if the field was never changed,
it will be equal to False and appear as 0 on the form view, leading to an error in the retrieval
of the appropriate IS Code.

We also improve the recompute whole sheet mecanism, applying changes to salary attachments and
contract modifications.

closes odoo/enterprise#65799

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>