[FIX] planning: add role on employee and shift in test case
Before this commit, the test `test_auto_plan_employee_with_break_company_no_breaks`
fails because the auto plan feature will plan the open shifts to a
resource generated by appointment feature since those resources are
always available according to their calendar.

This commit adds a role on the open shift and the employee created in
that test to be sure we will just fetch the resources with the role of
the open shift since the test just makes sure the resource doing night
shifts with breaks is correctly assigned to that shift.

runbot-115194

task-4679959

closes odoo/enterprise#82394

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>