[FIX] l10n_mx_reports_closing,account_reports: display Month 13
The display for the Month 13 date filter was kind of ugly after the
date filters refactoring. This makes it more useable and pretty to use.
It doesn't need to select the year at the same time as the Month 13
filter (they are now exclusive) with a choice to select the year into
the Month 13 filter.

closes odoo/enterprise#68623

X-original-commit: 014cb7783b1e2772c77455db216172944cc039a8
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>