[IMP] account_accountant: Created bank reconciliation action record
Created an XML record for the bank reconciliation action.
The created record was used in '_action_open_bank_reconciliation_widget' instead of a dictionary action.

Before this commit, the action state, which was not entirely encoded in the new URL,
would be erased on refresh or on opening the reconciliation link in a new tab.
Currently, the action state is stored in an XML record to avoid this problem.

task-4010694

closes odoo/enterprise#65843

Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>