[FIX] l10n_pe_edi: Fix-of-fix: allow IAP calls in external tests
PR #62452 (15.0+) added extra IAP calls that retry requesting the
cancellation status in case the cancellation was not processed yet.

In odoo/odoo#122663 (17.0+), IAP calls were disabled in tests.

When fw-porting the first PR above, this change of behaviour in 17.0
was not taken into account meaning that the tests kept failing in 17.0+.

We fix this by applying the same strategy as #62452 - patch the thread
to remove the testing attribute during IAP calls.

runbot-64793

closes odoo/enterprise#65926

X-original-commit: 6f1ec0b34b11477fd6aaf84f71ea0aeda2296ce3
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>