[FIX] timesheet_grid: verify timesheet existence in onchange method
Versions
--------
- 17.0
- saas-17.1
- saas-17.2
- saas-17.3

Steps
-----
1. Go to timesheets;
2. go to list view;
3. edit a timesheet line.

Issue
-----
`UncaughtPromiseError`

Cause
-----
Commit bae2e8146d48 added the `onRecordChanged` method, which fixed the
issue of the timer not recording time spent on a project.

The issue is that this method is also called when changing a row from
the list view, and there's no `timesheet` value set by the timer,
causing an error when trying to access `this.timesheet.resId`.

Solution
--------
Check `this.timerState.timesheetId` instead of `this.timesheet.resId`.
`this.timerState` is guaranteed to exist, and its `timesheetId` value
will be `undefined` if the timer isn't running, skipping the
timer-specific logic.

Also handle default projects when set.

opw-4027099

closes odoo/enterprise#65879

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>