[FIX] planning: fix slot duplication on recurring shift
Currently, it is impossible to duplicate a recurring shift, because the
"Edit Recurring Shift" modal opens instead. This PR fixes it by
preventing this modal from opening and copying the shift as it should.

The problem is that the "dragPillDrop" method will open the modal if the
shift is recurring, no matter if you are trying to copy it or not. To
fix it, a condition was added to check if the shift is being copied or
not, and if so, it won't open the modal.

task-3978527

closes odoo/enterprise#66172

X-original-commit: 7abe200e7c441220223db9590eeb7b30c59d2fe1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>