[FIX] sale_subscription: compatibility fix for invoice_line name in Subscriptions
Description of the issue this commit addresses:

Since the redesign of the description column of the invoice lines, that column
is not supposed to contain the name of the product anymore. At the moment, the
computation of the line name still uses it in the subscription app.
Referenced commit is:
https://github.com/odoo/odoo/pull/152869/commits/216634fd048c568d5b398134ba948a301bf76e3c

Desired behavior after the commit is merged:

When creating an invoice for a subscription, the invoice lines' description only
contain the custom description of the product and not the name of the product.

task-4013483

closes odoo/enterprise#65934

X-original-commit: 742dc32c442c2ded899a4072c7e33165a5f4aa35
Related: odoo/odoo#171677
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>