[FIX] data_cleaning: make the `_clean_records` deterministic
The `test_cleaning_action_case_lower` expects a
search result to be in a certain order, but the
order is not specified in the query done in
`_clean_records` and then the result is different.
In postgresql 16 certainly something has changed
in the planner that makes the result of
`_clean_records` undeterministic.

closes odoo/enterprise#66034

X-original-commit: f5171c3aa443ab51ab56e2acaf98be6f351c1fa7
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>