[IMP] *: load lazy assets in studio
With https://github.com/odoo/odoo/pull/171073, a system that allows to
lazy load views is introduced. For simplicity, we make studio load the
bundle "web.assets_backend_lazy" in order to be able to easily define
the editors for all views (lazy or not). Note that since the lazy bundle
is already loaded when launching unit tests, we avoid to load it twice
(hence the new bundle "web_studio.studio_assets_minimal").

Task ID: `3546321`

Part-of: odoo/enterprise#65641
Related: odoo/odoo#171073
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>