[IMP] social: handle deleted posts
This commit adds a new error modal when you clicking on deleted tweets and
posts.
It was decided to do so in stable as properly handling deleted posts requires
a more consistent approach for each social network.
This work should instead be done in later versions as such work will require a
lot of changes. Which is against our stable version policy.

task-3999571

closes odoo/enterprise#73214

X-original-commit: 81bf23a9863a221372d4a51895972be89b05de1f
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>