[IMP] social(_twitter): handle deleted tweets
This commit adds a new error modal when you clicking on deleted tweets and
posts.
It was decided to do so in stable as properly handling deleted posts requires
a more consistent approach for each social network.
This work should instead be done in later versions as such work will require a
lot of changes. Which is against our stable version policy.

task-3999571

closes odoo/enterprise#72934

X-original-commit: a17c8dcc8e18cd78727f5f1b2eacc4e5399a350e
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>