[FIX] documents_project: fix not-null constrains error on install w/ demo
To reproduce:
- Create a new db with demo data and install `sale_timesheet` module
- Then as a second step, install `documents_project`

The installation will crash with:
```
psycopg2.errors.NotNullViolation: null value in column "folder_id" of
relation "documents_document" violates not-null constraint
```

When using `search` with no corresponding field, `convert_xml()` will
only consider the first value, and in our case because of the ordering
of project model, we're not going to setup the document folder on
project which is not `project.project_project_1`.

This commit force creating documents folders for all projects with
`use_documents` option set.

closes odoo/enterprise#65964

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>