[FIX] web_enterprise: fix Studio app icon size
This commit fixes the size of studio icons when the window is resized. The w-75
class was applied, which caused the element to be resized instead of keeping the
desired width of 70px, like other icons do.

Now, the icon keep the same dimensions, like any other app icon.

task-3998287

closes odoo/enterprise#66123

Signed-off-by: Pierre Paridans (app) <app@odoo.com>