[FIX] sign: resize item after page change
Before this commit, when you drag and drop a resizeable sign item from one page
to another, it was not possible to resize the item: it was buggy because the
updated sign item changed the page and the resize function had the old object
binded to it.

After this commit, it is possible resizing the sign items when clicking the
border without any problems, from any page. This was obtained by re-binding the
updated sign item to the resize function.

task-4032512

closes odoo/enterprise#66676

X-original-commit: 31018ccee94f8647db025af6703636a92b90c84e
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>