[FIX] iot: notify right ws channel on printer status
Before this commit:
If an IoT is set on a company (other than the default one), when
printing a report set to an IoT device. It would always show the
"error to print" pop-up.

This happened as the notification that the printing operation was
successful was always send to the channel of the first company.

After this commit:
We force to use the company set on the device of the identifier
(which is the iot box company).

This heuristic is not perfect as different IoT might have devices
with the same identifier, in this case, it would accidentally send
to the wrong IoT channel.
To solve this other issue, another PR in the IoT will force to send
the IoT mac address in order to make sure to use the right IoT. As
the code of the IoT might be outdated, we define it with a default
value.

opw-3965623

closes odoo/enterprise#66294

X-original-commit: baeabeca59660091e86ae8ae51a040971b7cff82
Related: odoo/odoo#172384
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Loan Sens (lse) <lse@odoo.com>