[ADD] l10n_mx_edi_{40,stock_extended_{30,31}}: support version 3.1 of the delivery guide
This will add support for the new version 3.1 of the Mexian Carta Porte (Delivery Guide).

The main change is that the Many2one `l10n_mx_edi_customs_regime_id` field on `stock.picking` becomes a Many2many.
The existing ids are transferred to the new join table in a `_post_init_hook()`.

task-4000227

closes odoo/enterprise#66623

X-original-commit: 0c89f4e06271e28211b3f510670cc0a7991e53d4
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>