[FIX] planning: use employee tz for open slot datetimes
Versions
--------
- 16.0+

Steps
-----
1. Create an open slot in a planning;
2. publish & send an email to an employee who could fill it;
3. open the planning link sent via email;
4. open the same URL while logged out.

Issue
-----
The datetimes of the open slot are displayed in UTC.

Cause
-----
Commit cf256e94d333 modified the planning template to use the `datetime`
widget to format the dates instead of the `format_datetime` function
provided by the `ShiftController`. Doing so, it lost access to the
`employee_tz` value, derived from the email recipient, and hence
defaulted to UTC.

Solution
--------
Format the dates using `format_datetime`.

opw-3967787

closes odoo/enterprise#66284

X-original-commit: 4b8a8e8fbedd0f7cc35b1f0c9872a9ff241714eb
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>