[IMP] mail: move seen indicator into the bubble
Before this PR the position of the message seen indicator doesn't feel
consistent: if the message is squashed, its position is on the left. Otherwise,
it's on the right after the date of the message.

This PR removed those 2 positions and moved the message seen indicators on the
bottom right of the message bubble.

closes odoo/enterprise#66025

Related: odoo/odoo#171767
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>