[FIX] planning: fix sticky toast notification
Version:
saas-17.1

Before this commit:
the notification for the auto plan would stay on the screen even when navigating
to another page.

cause:
The notification were set to be sticky but did not properly handle removing the
notification when navigate to another page.

Solution:
added a method to remove notifications after navigating to another page.

task-4023197

closes odoo/enterprise#68092

X-original-commit: 3c88bbc82616ea14f803bccfe7de218f405650cf
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>