[FIX] project_enterprise: fix auto plan warning notification
Version:
saas-17.2

Before this commit:
the warning notification for the auto plan would stay on the screen even when
navigating to another page.

cause:
the warning notification were set to be sticky but did not properly handle
removing the notification when navigate to another page.

Solution:
added a method to remove notifications after navigating to another page.

Test case is added.

task-4023197

closes odoo/enterprise#68105

X-original-commit: 813f5e6f0ab5da256b0f74e59f44a2d5d6cae227
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>