[FIX] whatsapp: avoid double-sending the same body
When sending templates en masse to multiple recipients
they sometimes may have the same number.

If their numbers are the same, and the template resolves to the same value
it is useless to send them the message multiple times.
This behavior could be viewed as spam.

Now if all template variables resolve to the same value, messages will be
cancelled immediately. Similarly to what is done for emails.

task-4029026

Part-of: odoo/enterprise#78548
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>