[REF] mrp_mps: calculate mps_sequence
Rework how the sequence of each MPS record is calculated.

Before this PR: when you add a new record with a bom, records
for the components of the bom are automatically created.
The sequence of a record depends on if it is added manually
or automatically:
- manual -> sequence = 10
- auto -> sequence = parent.sequence + 1
Issues:
- it doesn't work for multi-level boms because only the first
level of components is added automatically.
- it is impossible to add new records _above_ existing ones since
sequence are not recomputed.

After this PR: when new records are created, Odoo will fetch once all
existing records linked to the new one (components and final products)
via indirect_demand_tree and re-assign all sequences. This solves both issues:
- no matter how a record is created or in which order, its sequence will
be based on the indirect_demand_tree
- if you add a new final product where a previous final_product is now a
component, it will correctly put the new final product above the old one.

task-4033339

Part-of: odoo/enterprise#65989
Related: odoo/upgrade#6241
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>