[FIX] quality_mrp: move `_update_lot_from_lot_line` to fix dependencies
Steps to reproduce:

- Install mrp_quality and uninstall mrp_workrder
- Create products A, B, and C
    - Product A: Storable, Manufacturing:
        - BOM: - 1 x product B
    - Product B: Storable,  tracked by Lot, Manufacturing route with:
        - BOM: - 1 x Product C
        - Quality check: Op manufacturing and Control Per Quantity
    - Product C: Storable, On hand quantity of 10
- Create MO for product B and complete it with a Lot id
- Create MO for product A > confirm

> Traceback error

Cause of the issue:

The `move_id` field of the `quality.check` model is used in the
`_update_lot_from_lot_line` method of that same model. However, the
method is defined in the `quality_mrp` module while the `move_id` field
is defined in the `mrp_workorder` module.

Since `quality_mrp` depends on `mrp` and since `mrp_workorder` is auto
installed when `mrp` is installed, this should not be an issue. However,
since `mrp_workorder` is not a strict dependency of the `quality_mrp`
module, the later in not uninstalled when `mrp_workorder` is. As such,
it is possible to not have a `move_id` field on the `quality.check`
model while the filed is used in the call of the
`_update_lot_from_lot_line` mdethod defined in `quality_mrp`.

For the posterity:

The method was introduced by commit 8ab6c56 in saas-16.4 for a fix and
was later an later updated by commit 801ff42 for an other fix.

opw-3994004

closes odoo/enterprise#66063

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>