[IMP] *: Properly converts py to JS locales
Serbian is a rare example of "synchronic digraphia", meaning that it has
two writing systems that coexist and are used interchangeably by its
speakers: Cyrillic and Latin.

To handle this feature of Serbian in Odoo, we use a locale with a
special modifier to specify the script to use: sr@latin.

Problem: This locale is not recognized by the JavaScript APIs that
implement the BCP 47 format, which leads to errors when trying to use
them with the sr@latin locale.

This commit adapts the code to use the properly formatted locale, thus
avoiding unexpected crashes.

*: sale_timesheet_enterprise, spreadsheet_edition, web_studio

Task-4014022

closes odoo/enterprise#66331

X-original-commit: fc058db80bfd4947bfcc6057da73240e974ff568
Related: odoo/odoo#172453
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>