[IMP] spreadsheet_edition: JS APIs support for sr@latin
Serbian is a rare example of "synchronic digraphia", meaning that it has
two writing systems that coexist and are used interchangeably by its
speakers: Cyrillic and Latin.

To handle this feature of Serbian in Odoo, we use a locale with a
special modifier to specify the script to use: sr@latin.

Problem: This locale is not recognized by the JavaScript APIs that
implement the BCP 47 format, which leads to errors when trying to use
them with the sr@latin locale.

This PR provides a helper to convert the locales from the Python side
for use on the JavaScript side.

Task-4014022

closes odoo/enterprise#66061

Related: odoo/odoo#171837
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>