[FIX] account_asset: compute residual_value in wizard
When the date in the modify wizard for assets is changed, the
residual_value should be updated to give a better view to the user.

As this value is changed, the tests had to be changed accordingly.

The gain or loss account is now correctly adapted from the computed
value (itself computed on the date) and the Asset Counterpart Account is
hidden if there is no gain.

task-3981380

closes odoo/enterprise#72798

X-original-commit: d606413d40162db5b2b01ac0d79b3369e75c7d98
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>