[FIX] sale_subscription: fix subscription portal view
Changes
-----
Since 17.2, the subscription view in the portal only shows what will be invoiced
next. SO lines of non recurring products are hidden if already invoiced. This
commit makes the following changes:

1. When viewing a subscription in the portal from the Orders button, the
breadcrumb links to the Subscriptions page instead of the Sales Orders page.
2. To allow the user to see non recurring products of a subscription, accessing
previous orders from the "History" link of a subscription will link to the sales
orders (which show the non recurring lines).

opw-4028536

closes odoo/enterprise#69116

X-original-commit: a67e2b7fe1c38c06509e96725cab3f23a7d79a1c
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>