[IMP] hr_contract_salary: correctly update employee info
Prior, employee information - job_title, job_id and department_id
were updated as soon as employee signed a new offer.
This is flawed behavior, as even though employee signed offer,
the new contract might have not started yet. In this task we correct this.

After this, above mentioned employee information will not be updated
as soon as offer is signed.

Moreover, functionally it makes sense to link originating offer with
the contract that was created from it.
Thus, we add originated_offer_id field.

task-3035630

closes odoo/enterprise#66073

Related: odoo/odoo#171967
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>