[FIX] l10n_nl_reports_sbr{_icp|_status_info}: fix accountant access rights
The setup of the password from a related field would throw errors
to user with accountant rights but no administration (settings) rights.
This showed other access rights problems with the SBR implementation.

The way we want to work with this is: an accountant without settings
rights can set up the password through the wizard if no password is
stored on the company (for databases that didn't start SBR with
the status info module). Once the password is set and is correct (it is
the right one for the certificate), it shouldn't ask the user to put it
again and is thus hidden in the wizard.

ticket-4024344

closes odoo/enterprise#66762

X-original-commit: 00e994a9f5b2992f680afd316ea76c7cb5a467bb
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>