[FIX] hr_contract_salary: disable dependent benefits
When we are loading an offer in the salary configurator, if values are
set to a benefit that depends on another one which is disabled. You are
keeping the value of the contract, and it is not disabled as it should
because the mandatory benefit is not set.

We've got this issue because the variable in the loop has the same name
as the one we really need declared before ('mandatoryBenefitSelected').

We also remove the loop and the split, as it is already done in the
function 'updateDependentBenefits' that is called.

closes odoo/enterprise#66182

Task-id: NONE
X-original-commit: 21195e295aed481db6355a492034a8cf585e562b
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>